Logic for Half Days Implemented
Updated ints to doubles
Checks for single days that are also half days
Added check for if there are enough holidays before creating or updating event
Updated Tests
Updated Schema

+++ v2/AdminCore.DatabaseMigration/db/migrations/V1.0__original_schema.sql
@@ -139,7 +139,7 @@ CREATE TABLE IF NOT EXISTS public.employee
     password character varying(255) COLLATE pg_catalog."default",
     start_date date NOT NULL,
     surname character varying(255) COLLATE pg_catalog."default",
-    total_holidays integer NOT NULL,
+    total_holidays decimal NOT NULL,
     CONSTRAINT employee_pkey PRIMARY KEY (employee_id),
     CONSTRAINT employee_country_id_fkey FOREIGN KEY (country_id)
         REFERENCES public.country (country_id) MATCH SIMPLE
